fix: fix PR issues
